C - Function pointers

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,5 @@
 #include "function_pointers.h"
-
+#include<stddef.h>
 /**
  * array_iterator - function that executes a function given as a
  *  parameter on each element of an array.
@@ -11,7 +11,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	if (array != NULL && action != NULL)
 	{
